+++ packages/core/src/render3/index.ts
@@ -19,7 +19,6 @@ export {RenderFlags} from './interfaces/definition';
 export {CssSelectorList} from './interfaces/projection';
 
 
-
 // clang-format off
 export {
   allocHostVars,
