+++ packages/core/src/render3/instructions/namespace.ts
@@ -1,0 +1,9 @@
+/**
+ * @license
+ * Copyright Google Inc. All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+export {namespaceHTML, namespaceMathML, namespaceSVG} from '../state';
