+++ packages/core/src/render3/instructions/styling_instructions.ts
@@ -18,8 +18,7 @@ import {renderStringify} from '../util/misc_utils';
 import {getRootContext} from '../util/view_traversal_utils';
 import {getTNode} from '../util/view_utils';
 
-import {scheduleTick} from './instructions';
-import {setInputsForProperty} from './shared';
+import {scheduleTick, setInputsForProperty} from './shared';
