+++ packages/core/src/render3/players.ts
@@ -8,11 +8,12 @@
 import '../util/ng_dev_mode';
 
 import {getLContext} from './context_discovery';
-import {scheduleTick} from './instructions/all';
+import {scheduleTick} from './instructions/shared';
 import {ComponentInstance, DirectiveInstance, Player} from './interfaces/player';
 import {RootContextFlags} from './interfaces/view';
 import {addPlayerInternal, getOrCreatePlayerContext, getPlayerContext, getPlayersInternal, getStylingContext, throwInvalidRefError} from './styling/util';
 import {getRootContext} from './util/view_traversal_utils';
+
 
 /**
  * Adds a player to an element, directive or component instance that will later be
