+++ packages/core/test/bundling/cyclic_import/bundle.golden_symbols.json
@@ -378,6 +378,9 @@
     "name": "getNameOnlyMarkerIndex"
   },
   {
+    "name": "getNamespace"
+  },
+  {
     "name": "getNativeAnchorNode"
   },
   {
